Grabbing label in queries and adding URL in xsl
Was using the URI of the dcoid but this causes problems because it starts with a http: instead of https:. When clicking on the dco-id and you were logged in
it would log you out and force you to log back in (not chrome, chrome did the right thing). So grabbing the label now instead of the URI of the dcoid and
tacking on https://dx.deepcarbon.net/

Also changed a few xsl to use relative paths to the summary pages instead of full paths to make working in test environments easier.